Create a monitor file that can be consumed by system-image
This file, whose contents are in the format
timestamp-branch-buildNumber-commitHash, is used by system-image to
determine if the image it's about to download has changed. system-image
also embeds the contents of the monitor file into the deployed image's
version information.

Old monitor files for xenial are in the format date-$RANDOM. This seems better.